only break split statement for the ease of review
