+++ src/Neo.Compiler.CSharp/MethodConvert/MethodConvert.cs
@@ -33,7 +33,7 @@ using System.Text;
 
 namespace Neo.Compiler
 {
-    class MethodConvert
+    partial class MethodConvert
     {
 
         #region Fields
@@ -779,619 +779,6 @@ namespace Neo.Compiler
                 default:
                     throw new CompilationException(statement, DiagnosticId.SyntaxNotSupported, $"Unsupported syntax: {statement}");
             }
-        }
-
-        private void ConvertBlockStatement(SemanticModel model, BlockSyntax syntax)
-        {
-            _blockSymbols.Push(new List<ILocalSymbol>());
-            using (InsertSequencePoint(syntax.OpenBraceToken))
-                AddInstruction(OpCode.NOP);
-            foreach (StatementSyntax child in syntax.Statements)
-                ConvertStatement(model, child);
-            using (InsertSequencePoint(syntax.CloseBraceToken))
-                AddInstruction(OpCode.NOP);
-            foreach (ILocalSymbol symbol in _blockSymbols.Pop())
-                RemoveLocalVariable(symbol);
-        }
-
-        private void ConvertBreakStatement(BreakStatementSyntax syntax)
-        {
-            using (InsertSequencePoint(syntax))
-                if (_tryStack.TryPeek(out ExceptionHandling? result) && result.BreakTargetCount == 0)
-                    Jump(OpCode.ENDTRY_L, _breakTargets.Peek());
-                else
-                    Jump(OpCode.JMP_L, _breakTargets.Peek());
-        }
-
-        private void ConvertCheckedStatement(SemanticModel model, CheckedStatementSyntax syntax)
-        {
-            _checkedStack.Push(syntax.Keyword.IsKind(SyntaxKind.CheckedKeyword));
-            ConvertBlockStatement(model, syntax.Block);
-            _checkedStack.Pop();
-        }
-
-        private void ConvertContinueStatement(ContinueStatementSyntax syntax)
-        {
-            using (InsertSequencePoint(syntax))
-                if (_tryStack.TryPeek(out ExceptionHandling? result) && result.ContinueTargetCount == 0)
-                    Jump(OpCode.ENDTRY_L, _continueTargets.Peek());
-                else
-                    Jump(OpCode.JMP_L, _continueTargets.Peek());
-        }
-
-        private void ConvertDoStatement(SemanticModel model, DoStatementSyntax syntax)
-        {
-            JumpTarget startTarget = new();
-            JumpTarget continueTarget = new();
-            JumpTarget breakTarget = new();
-            PushContinueTarget(continueTarget);
-            PushBreakTarget(breakTarget);
-            startTarget.Instruction = AddInstruction(OpCode.NOP);
-            ConvertStatement(model, syntax.Statement);
-            continueTarget.Instruction = AddInstruction(OpCode.NOP);
-            using (InsertSequencePoint(syntax.Condition))
-                ConvertExpression(model, syntax.Condition);
-            Jump(OpCode.JMPIF_L, startTarget);
-            breakTarget.Instruction = AddInstruction(OpCode.NOP);
-            PopContinueTarget();
-            PopBreakTarget();
-        }
-
-        private void ConvertEmptyStatement(EmptyStatementSyntax syntax)
-        {
-            using (InsertSequencePoint(syntax))
-                AddInstruction(OpCode.NOP);
-        }
-
-        private void ConvertExpressionStatement(SemanticModel model, ExpressionStatementSyntax syntax)
-        {
-            ITypeSymbol type = model.GetTypeInfo(syntax.Expression).Type!;
-            using (InsertSequencePoint(syntax))
-            {
-                ConvertExpression(model, syntax.Expression);
-                if (type.SpecialType != SpecialType.System_Void)
-                    AddInstruction(OpCode.DROP);
-            }
-        }
-
-        private void ConvertForEachStatement(SemanticModel model, ForEachStatementSyntax syntax)
-        {
-            ITypeSymbol type = model.GetTypeInfo(syntax.Expression).Type!;
-            if (type.Name == "Iterator")
-            {
-                ConvertIteratorForEachStatement(model, syntax);
-            }
-            else
-            {
-                ConvertArrayForEachStatement(model, syntax);
-            }
-        }
-
-        private void ConvertIteratorForEachStatement(SemanticModel model, ForEachStatementSyntax syntax)
-        {
-            ILocalSymbol elementSymbol = model.GetDeclaredSymbol(syntax)!;
-            JumpTarget startTarget = new();
-            JumpTarget continueTarget = new();
-            JumpTarget breakTarget = new();
-            byte iteratorIndex = AddAnonymousVariable();
-            byte elementIndex = AddLocalVariable(elementSymbol);
-            PushContinueTarget(continueTarget);
-            PushBreakTarget(breakTarget);
-            using (InsertSequencePoint(syntax.ForEachKeyword))
-            {
-                ConvertExpression(model, syntax.Expression);
-                AccessSlot(OpCode.STLOC, iteratorIndex);
-                Jump(OpCode.JMP_L, continueTarget);
-            }
-            using (InsertSequencePoint(syntax.Identifier))
-            {
-                startTarget.Instruction = AccessSlot(OpCode.LDLOC, iteratorIndex);
-                Call(ApplicationEngine.System_Iterator_Value);
-                AccessSlot(OpCode.STLOC, elementIndex);
-            }
-            ConvertStatement(model, syntax.Statement);
-            using (InsertSequencePoint(syntax.Expression))
-            {
-                continueTarget.Instruction = AccessSlot(OpCode.LDLOC, iteratorIndex);
-                Call(ApplicationEngine.System_Iterator_Next);
-                Jump(OpCode.JMPIF_L, startTarget);
-            }
-            breakTarget.Instruction = AddInstruction(OpCode.NOP);
-            RemoveAnonymousVariable(iteratorIndex);
-            RemoveLocalVariable(elementSymbol);
-            PopContinueTarget();
-            PopBreakTarget();
-        }
-
-        private void ConvertArrayForEachStatement(SemanticModel model, ForEachStatementSyntax syntax)
-        {
-            ILocalSymbol elementSymbol = model.GetDeclaredSymbol(syntax)!;
-            JumpTarget startTarget = new();
-            JumpTarget continueTarget = new();
-            JumpTarget conditionTarget = new();
-            JumpTarget breakTarget = new();
-            byte arrayIndex = AddAnonymousVariable();
-            byte lengthIndex = AddAnonymousVariable();
-            byte iIndex = AddAnonymousVariable();
-            byte elementIndex = AddLocalVariable(elementSymbol);
-            PushContinueTarget(continueTarget);
-            PushBreakTarget(breakTarget);
-            using (InsertSequencePoint(syntax.ForEachKeyword))
-            {
-                ConvertExpression(model, syntax.Expression);
-                AddInstruction(OpCode.DUP);
-                AccessSlot(OpCode.STLOC, arrayIndex);
-                AddInstruction(OpCode.SIZE);
-                AccessSlot(OpCode.STLOC, lengthIndex);
-                Push(0);
-                AccessSlot(OpCode.STLOC, iIndex);
-                Jump(OpCode.JMP_L, conditionTarget);
-            }
-            using (InsertSequencePoint(syntax.Identifier))
-            {
-                startTarget.Instruction = AccessSlot(OpCode.LDLOC, arrayIndex);
-                AccessSlot(OpCode.LDLOC, iIndex);
-                AddInstruction(OpCode.PICKITEM);
-                AccessSlot(OpCode.STLOC, elementIndex);
-            }
-            ConvertStatement(model, syntax.Statement);
-            using (InsertSequencePoint(syntax.Expression))
-            {
-                continueTarget.Instruction = AccessSlot(OpCode.LDLOC, iIndex);
-                AddInstruction(OpCode.INC);
-                AccessSlot(OpCode.STLOC, iIndex);
-                conditionTarget.Instruction = AccessSlot(OpCode.LDLOC, iIndex);
-                AccessSlot(OpCode.LDLOC, lengthIndex);
-                Jump(OpCode.JMPLT_L, startTarget);
-            }
-            breakTarget.Instruction = AddInstruction(OpCode.NOP);
-            RemoveAnonymousVariable(arrayIndex);
-            RemoveAnonymousVariable(lengthIndex);
-            RemoveAnonymousVariable(iIndex);
-            RemoveLocalVariable(elementSymbol);
-            PopContinueTarget();
-            PopBreakTarget();
-        }
-
-        private void ConvertForEachVariableStatement(SemanticModel model, ForEachVariableStatementSyntax syntax)
-        {
-            ITypeSymbol type = model.GetTypeInfo(syntax.Expression).Type!;
-            if (type.Name == "Iterator")
-            {
-                ConvertIteratorForEachVariableStatement(model, syntax);
-            }
-            else
-            {
-                ConvertArrayForEachVariableStatement(model, syntax);
-            }
-        }
-
-        private void ConvertIteratorForEachVariableStatement(SemanticModel model, ForEachVariableStatementSyntax syntax)
-        {
-            ILocalSymbol[] symbols = ((ParenthesizedVariableDesignationSyntax)((DeclarationExpressionSyntax)syntax.Variable).Designation).Variables.Select(p => (ILocalSymbol)model.GetDeclaredSymbol(p)!).ToArray();
-            JumpTarget startTarget = new();
-            JumpTarget continueTarget = new();
-            JumpTarget breakTarget = new();
-            byte iteratorIndex = AddAnonymousVariable();
-            PushContinueTarget(continueTarget);
-            PushBreakTarget(breakTarget);
-            using (InsertSequencePoint(syntax.ForEachKeyword))
-            {
-                ConvertExpression(model, syntax.Expression);
-                AccessSlot(OpCode.STLOC, iteratorIndex);
-                Jump(OpCode.JMP_L, continueTarget);
-            }
-            using (InsertSequencePoint(syntax.Variable))
-            {
-                startTarget.Instruction = AccessSlot(OpCode.LDLOC, iteratorIndex);
-                Call(ApplicationEngine.System_Iterator_Value);
-                AddInstruction(OpCode.UNPACK);
-                AddInstruction(OpCode.DROP);
-                for (int i = 0; i < symbols.Length; i++)
-                {
-                    if (symbols[i] is null)
-                    {
-                        AddInstruction(OpCode.DROP);
-                    }
-                    else
-                    {
-                        byte variableIndex = AddLocalVariable(symbols[i]);
-                        AccessSlot(OpCode.STLOC, variableIndex);
-                    }
-                }
-            }
-            ConvertStatement(model, syntax.Statement);
-            using (InsertSequencePoint(syntax.Expression))
-            {
-                continueTarget.Instruction = AccessSlot(OpCode.LDLOC, iteratorIndex);
-                Call(ApplicationEngine.System_Iterator_Next);
-                Jump(OpCode.JMPIF_L, startTarget);
-            }
-            breakTarget.Instruction = AddInstruction(OpCode.NOP);
-            RemoveAnonymousVariable(iteratorIndex);
-            foreach (ILocalSymbol symbol in symbols)
-                if (symbol is not null)
-                    RemoveLocalVariable(symbol);
-            PopContinueTarget();
-            PopBreakTarget();
-        }
-
-        private void ConvertArrayForEachVariableStatement(SemanticModel model, ForEachVariableStatementSyntax syntax)
-        {
-            ILocalSymbol[] symbols = ((ParenthesizedVariableDesignationSyntax)((DeclarationExpressionSyntax)syntax.Variable).Designation).Variables.Select(p => (ILocalSymbol)model.GetDeclaredSymbol(p)!).ToArray();
-            JumpTarget startTarget = new();
-            JumpTarget continueTarget = new();
-            JumpTarget conditionTarget = new();
-            JumpTarget breakTarget = new();
-            byte arrayIndex = AddAnonymousVariable();
-            byte lengthIndex = AddAnonymousVariable();
-            byte iIndex = AddAnonymousVariable();
-            PushContinueTarget(continueTarget);
-            PushBreakTarget(breakTarget);
-            using (InsertSequencePoint(syntax.ForEachKeyword))
-            {
-                ConvertExpression(model, syntax.Expression);
-                AddInstruction(OpCode.DUP);
-                AccessSlot(OpCode.STLOC, arrayIndex);
-                AddInstruction(OpCode.SIZE);
-                AccessSlot(OpCode.STLOC, lengthIndex);
-                Push(0);
-                AccessSlot(OpCode.STLOC, iIndex);
-                Jump(OpCode.JMP_L, conditionTarget);
-            }
-            using (InsertSequencePoint(syntax.Variable))
-            {
-                startTarget.Instruction = AccessSlot(OpCode.LDLOC, arrayIndex);
-                AccessSlot(OpCode.LDLOC, iIndex);
-                AddInstruction(OpCode.PICKITEM);
-                AddInstruction(OpCode.UNPACK);
-                AddInstruction(OpCode.DROP);
-                for (int i = 0; i < symbols.Length; i++)
-                {
-                    if (symbols[i] is null)
-                    {
-                        AddInstruction(OpCode.DROP);
-                    }
-                    else
-                    {
-                        byte variableIndex = AddLocalVariable(symbols[i]);
-                        AccessSlot(OpCode.STLOC, variableIndex);
-                    }
-                }
-            }
-            ConvertStatement(model, syntax.Statement);
-            using (InsertSequencePoint(syntax.Expression))
-            {
-                continueTarget.Instruction = AccessSlot(OpCode.LDLOC, iIndex);
-                AddInstruction(OpCode.INC);
-                AccessSlot(OpCode.STLOC, iIndex);
-                conditionTarget.Instruction = AccessSlot(OpCode.LDLOC, iIndex);
-                AccessSlot(OpCode.LDLOC, lengthIndex);
-                Jump(OpCode.JMPLT_L, startTarget);
-            }
-            breakTarget.Instruction = AddInstruction(OpCode.NOP);
-            RemoveAnonymousVariable(arrayIndex);
-            RemoveAnonymousVariable(lengthIndex);
-            RemoveAnonymousVariable(iIndex);
-            foreach (ILocalSymbol symbol in symbols)
-                if (symbol is not null)
-                    RemoveLocalVariable(symbol);
-            PopContinueTarget();
-            PopBreakTarget();
-        }
-
-        private void ConvertForStatement(SemanticModel model, ForStatementSyntax syntax)
-        {
-            var variables = (syntax.Declaration?.Variables ?? Enumerable.Empty<VariableDeclaratorSyntax>())
-                .Select(p => (p, (ILocalSymbol)model.GetDeclaredSymbol(p)!))
-                .ToArray();
-            foreach (ExpressionSyntax expression in syntax.Initializers)
-                using (InsertSequencePoint(expression))
-                {
-                    ITypeSymbol type = model.GetTypeInfo(expression).Type!;
-                    ConvertExpression(model, expression);
-                    if (type.SpecialType != SpecialType.System_Void)
-                        AddInstruction(OpCode.DROP);
-                }
-            JumpTarget startTarget = new();
-            JumpTarget continueTarget = new();
-            JumpTarget conditionTarget = new();
-            JumpTarget breakTarget = new();
-            PushContinueTarget(continueTarget);
-            PushBreakTarget(breakTarget);
-            foreach (var (variable, symbol) in variables)
-            {
-                byte variableIndex = AddLocalVariable(symbol);
-                if (variable.Initializer is not null)
-                    using (InsertSequencePoint(variable))
-                    {
-                        ConvertExpression(model, variable.Initializer.Value);
-                        AccessSlot(OpCode.STLOC, variableIndex);
-                    }
-            }
-            Jump(OpCode.JMP_L, conditionTarget);
-            startTarget.Instruction = AddInstruction(OpCode.NOP);
-            ConvertStatement(model, syntax.Statement);
-            continueTarget.Instruction = AddInstruction(OpCode.NOP);
-            foreach (ExpressionSyntax expression in syntax.Incrementors)
-                using (InsertSequencePoint(expression))
-                {
-                    ITypeSymbol type = model.GetTypeInfo(expression).Type!;
-                    ConvertExpression(model, expression);
-                    if (type.SpecialType != SpecialType.System_Void)
-                        AddInstruction(OpCode.DROP);
-                }
-            conditionTarget.Instruction = AddInstruction(OpCode.NOP);
-            if (syntax.Condition is null)
-            {
-                Jump(OpCode.JMP_L, startTarget);
-            }
-            else
-            {
-                using (InsertSequencePoint(syntax.Condition))
-                    ConvertExpression(model, syntax.Condition);
-                Jump(OpCode.JMPIF_L, startTarget);
-            }
-            breakTarget.Instruction = AddInstruction(OpCode.NOP);
-            foreach (var (_, symbol) in variables)
-                RemoveLocalVariable(symbol);
-            PopContinueTarget();
-            PopBreakTarget();
-        }
-
-        private void ConvertGotoStatement(SemanticModel model, GotoStatementSyntax syntax)
-        {
-            using (InsertSequencePoint(syntax))
-                if (syntax.CaseOrDefaultKeyword.IsKind(SyntaxKind.None))
-                {
-                    ILabelSymbol symbol = (ILabelSymbol)model.GetSymbolInfo(syntax.Expression!).Symbol!;
-                    JumpTarget target = AddLabel(symbol, false);
-                    if (_tryStack.TryPeek(out ExceptionHandling? result) && result.State != ExceptionHandlingState.Finally && !result.Labels.Contains(symbol))
-                        result.PendingGotoStatments.Add(Jump(OpCode.ENDTRY_L, target));
-                    else
-                        Jump(OpCode.JMP_L, target);
-                }
-                else
-                {
-                    var labels = _switchStack.Peek();
-                    JumpTarget target = default!;
-                    if (syntax.CaseOrDefaultKeyword.IsKind(SyntaxKind.DefaultKeyword))
-                    {
-                        target = labels.First(p => p.Item1 is DefaultSwitchLabelSyntax).Item2;
-                    }
-                    else
-                    {
-                        object? value = model.GetConstantValue(syntax.Expression!).Value;
-                        foreach (var (l, t) in labels)
-                        {
-                            if (l is not CaseSwitchLabelSyntax cl) continue;
-                            object? clValue = model.GetConstantValue(cl.Value).Value;
-                            if (value is null)
-                            {
-                                if (clValue is null)
-                                {
-                                    target = t;
-                                    break;
-                                }
-                            }
-                            else
-                            {
-                                if (value.Equals(clValue))
-                                {
-                                    target = t;
-                                    break;
-                                }
-                            }
-                        }
-                    }
-                    if (_tryStack.TryPeek(out ExceptionHandling? result) && result.SwitchCount == 0)
-                        Jump(OpCode.ENDTRY_L, target);
-                    else
-                        Jump(OpCode.JMP_L, target);
-                }
-        }
-
-        private void ConvertIfStatement(SemanticModel model, IfStatementSyntax syntax)
-        {
-            JumpTarget elseTarget = new();
-            using (InsertSequencePoint(syntax.Condition))
-                ConvertExpression(model, syntax.Condition);
-            Jump(OpCode.JMPIFNOT_L, elseTarget);
-            ConvertStatement(model, syntax.Statement);
-            if (syntax.Else is null)
-            {
-                elseTarget.Instruction = AddInstruction(OpCode.NOP);
-            }
-            else
-            {
-                JumpTarget endTarget = new();
-                Jump(OpCode.JMP_L, endTarget);
-                elseTarget.Instruction = AddInstruction(OpCode.NOP);
-                ConvertStatement(model, syntax.Else.Statement);
-                endTarget.Instruction = AddInstruction(OpCode.NOP);
-            }
-        }
-
-        private void ConvertLabeledStatement(SemanticModel model, LabeledStatementSyntax syntax)
-        {
-            ILabelSymbol symbol = model.GetDeclaredSymbol(syntax)!;
-            JumpTarget target = AddLabel(symbol, true);
-            if (_tryStack.TryPeek(out ExceptionHandling? result))
-                foreach (Instruction instruction in result.PendingGotoStatments)
-                    if (instruction.Target == target)
-                        instruction.OpCode = OpCode.JMP_L;
-            target.Instruction = AddInstruction(OpCode.NOP);
-            ConvertStatement(model, syntax.Statement);
-        }
-
-        private void ConvertLocalDeclarationStatement(SemanticModel model, LocalDeclarationStatementSyntax syntax)
-        {
-            if (syntax.IsConst) return;
-            foreach (VariableDeclaratorSyntax variable in syntax.Declaration.Variables)
-            {
-                ILocalSymbol symbol = (ILocalSymbol)model.GetDeclaredSymbol(variable)!;
-                byte variableIndex = AddLocalVariable(symbol);
-                if (variable.Initializer is not null)
-                    using (InsertSequencePoint(variable))
-                    {
-                        ConvertExpression(model, variable.Initializer.Value);
-                        AccessSlot(OpCode.STLOC, variableIndex);
-                    }
-            }
-        }
-
-        private void ConvertReturnStatement(SemanticModel model, ReturnStatementSyntax syntax)
-        {
-            using (InsertSequencePoint(syntax))
-            {
-                if (syntax.Expression is not null)
-                    ConvertExpression(model, syntax.Expression);
-                if (_tryStack.Count > 0)
-                    Jump(OpCode.ENDTRY_L, _returnTarget);
-                else
-                    Jump(OpCode.JMP_L, _returnTarget);
-            }
-        }
-
-        private void ConvertSwitchStatement(SemanticModel model, SwitchStatementSyntax syntax)
-        {
-            var sections = syntax.Sections.Select(p => (p.Labels, p.Statements, Target: new JumpTarget())).ToArray();
-            var labels = sections.SelectMany(p => p.Labels, (p, l) => (l, p.Target)).ToArray();
-            PushSwitchLabels(labels);
-            JumpTarget breakTarget = new();
-            byte anonymousIndex = AddAnonymousVariable();
-            PushBreakTarget(breakTarget);
-            using (InsertSequencePoint(syntax.Expression))
-            {
-                ConvertExpression(model, syntax.Expression);
-                AccessSlot(OpCode.STLOC, anonymousIndex);
-            }
-            foreach (var (label, target) in labels)
-            {
-                switch (label)
-                {
-                    case CasePatternSwitchLabelSyntax casePatternSwitchLabel:
-                        using (InsertSequencePoint(casePatternSwitchLabel))
-                        {
-                            JumpTarget endTarget = new();
-                            ConvertPattern(model, casePatternSwitchLabel.Pattern, anonymousIndex);
-                            Jump(OpCode.JMPIFNOT_L, endTarget);
-                            if (casePatternSwitchLabel.WhenClause is not null)
-                            {
-                                ConvertExpression(model, casePatternSwitchLabel.WhenClause.Condition);
-                                Jump(OpCode.JMPIFNOT_L, endTarget);
-                            }
-                            Jump(OpCode.JMP_L, target);
-                            endTarget.Instruction = AddInstruction(OpCode.NOP);
-                        }
-                        break;
-                    case CaseSwitchLabelSyntax caseSwitchLabel:
-                        using (InsertSequencePoint(caseSwitchLabel))
-                        {
-                            AccessSlot(OpCode.LDLOC, anonymousIndex);
-                            ConvertExpression(model, caseSwitchLabel.Value);
-                            AddInstruction(OpCode.EQUAL);
-                            Jump(OpCode.JMPIF_L, target);
-                        }
-                        break;
-                    case DefaultSwitchLabelSyntax defaultSwitchLabel:
-                        using (InsertSequencePoint(defaultSwitchLabel))
-                        {
-                            Jump(OpCode.JMP_L, target);
-                        }
-                        break;
-                    default:
-                        throw new CompilationException(label, DiagnosticId.SyntaxNotSupported, $"Unsupported syntax: {label}");
-                }
-            }
-            RemoveAnonymousVariable(anonymousIndex);
-            Jump(OpCode.JMP_L, breakTarget);
-            foreach (var (_, statements, target) in sections)
-            {
-                target.Instruction = AddInstruction(OpCode.NOP);
-                foreach (StatementSyntax statement in statements)
-                    ConvertStatement(model, statement);
-            }
-            breakTarget.Instruction = AddInstruction(OpCode.NOP);
-            PopSwitchLabels();
-            PopBreakTarget();
-        }
-
-        private void ConvertThrowStatement(SemanticModel model, ThrowStatementSyntax syntax)
-        {
-            using (InsertSequencePoint(syntax))
-                Throw(model, syntax.Expression);
-        }
-
-        private void ConvertTryStatement(SemanticModel model, TryStatementSyntax syntax)
-        {
-            JumpTarget catchTarget = new();
-            JumpTarget finallyTarget = new();
-            JumpTarget endTarget = new();
-            AddInstruction(new Instruction { OpCode = OpCode.TRY_L, Target = catchTarget, Target2 = finallyTarget });
-            _tryStack.Push(new ExceptionHandling { State = ExceptionHandlingState.Try });
-            ConvertStatement(model, syntax.Block);
-            Jump(OpCode.ENDTRY_L, endTarget);
-            if (syntax.Catches.Count > 1)
-                throw new CompilationException(syntax.Catches[1], DiagnosticId.MultiplyCatches, "Only support one single catch.");
-            if (syntax.Catches.Count > 0)
-            {
-                CatchClauseSyntax catchClause = syntax.Catches[0];
-                if (catchClause.Filter is not null)
-                    throw new CompilationException(catchClause.Filter, DiagnosticId.CatchFilter, $"Unsupported syntax: {catchClause.Filter}");
-                _tryStack.Peek().State = ExceptionHandlingState.Catch;
-                ILocalSymbol? exceptionSymbol = null;
-                byte exceptionIndex;
-                if (catchClause.Declaration is null)
-                {
-                    exceptionIndex = AddAnonymousVariable();
-                }
-                else
-                {
-                    exceptionSymbol = model.GetDeclaredSymbol(catchClause.Declaration);
-                    exceptionIndex = exceptionSymbol is null
-                        ? AddAnonymousVariable()
-                        : AddLocalVariable(exceptionSymbol);
-                }
-                using (InsertSequencePoint(catchClause.CatchKeyword))
-                    catchTarget.Instruction = AccessSlot(OpCode.STLOC, exceptionIndex);
-                _exceptionStack.Push(exceptionIndex);
-                ConvertStatement(model, catchClause.Block);
-                Jump(OpCode.ENDTRY_L, endTarget);
-                if (exceptionSymbol is null)
-                    RemoveAnonymousVariable(exceptionIndex);
-                else
-                    RemoveLocalVariable(exceptionSymbol);
-                _exceptionStack.Pop();
-            }
-            if (syntax.Finally is not null)
-            {
-                _tryStack.Peek().State = ExceptionHandlingState.Finally;
-                finallyTarget.Instruction = AddInstruction(OpCode.NOP);
-                ConvertStatement(model, syntax.Finally.Block);
-                AddInstruction(OpCode.ENDFINALLY);
-            }
-            endTarget.Instruction = AddInstruction(OpCode.NOP);
-            _tryStack.Pop();
-        }
-
-        private void ConvertWhileStatement(SemanticModel model, WhileStatementSyntax syntax)
-        {
-            JumpTarget continueTarget = new();
-            JumpTarget breakTarget = new();
-            PushContinueTarget(continueTarget);
-            PushBreakTarget(breakTarget);
-            continueTarget.Instruction = AddInstruction(OpCode.NOP);
-            using (InsertSequencePoint(syntax.Condition))
-            {
-                ConvertExpression(model, syntax.Condition);
-                Jump(OpCode.JMPIFNOT_L, breakTarget);
-            }
-            ConvertStatement(model, syntax.Statement);
-            Jump(OpCode.JMP_L, continueTarget);
-            breakTarget.Instruction = AddInstruction(OpCode.NOP);
-            PopContinueTarget();
-            PopBreakTarget();
         }
 
         #endregion
@@ -3558,8 +2945,6 @@ namespace Neo.Compiler
         }
 
         #endregion
-
-
 
         #region StackHelpers
