+++ src/Neo.Compiler.CSharp/MethodConvert/Statement/CheckedStatement.cs
@@ -1,0 +1,26 @@
+// Copyright (C) 2015-2023 The Neo Project.
+//
+// The Neo.Compiler.CSharp is free software distributed under the MIT
+// software license, see the accompanying file LICENSE in the main directory
+// of the project or http://www.opensource.org/licenses/mit-license.php
+// for more details.
+//
+// Redistribution and use in source and binary forms with or without
+// modifications are permitted.
+
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+
+namespace Neo.Compiler
+{
+    partial class MethodConvert
+    {
+        private void ConvertCheckedStatement(SemanticModel model, CheckedStatementSyntax syntax)
+        {
+            _checkedStack.Push(syntax.Keyword.IsKind(SyntaxKind.CheckedKeyword));
+            ConvertBlockStatement(model, syntax.Block);
+            _checkedStack.Pop();
+        }
+    }
+}
