+++ src/Neo.Compiler.CSharp/MethodConvert/Statement/DoStatement.cs
@@ -1,0 +1,37 @@
+// Copyright (C) 2015-2023 The Neo Project.
+//
+// The Neo.Compiler.CSharp is free software distributed under the MIT
+// software license, see the accompanying file LICENSE in the main directory
+// of the project or http://www.opensource.org/licenses/mit-license.php
+// for more details.
+//
+// Redistribution and use in source and binary forms with or without
+// modifications are permitted.
+
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Neo.VM;
+
+namespace Neo.Compiler
+{
+    partial class MethodConvert
+    {
+        private void ConvertDoStatement(SemanticModel model, DoStatementSyntax syntax)
+        {
+            JumpTarget startTarget = new();
+            JumpTarget continueTarget = new();
+            JumpTarget breakTarget = new();
+            PushContinueTarget(continueTarget);
+            PushBreakTarget(breakTarget);
+            startTarget.Instruction = AddInstruction(OpCode.NOP);
+            ConvertStatement(model, syntax.Statement);
+            continueTarget.Instruction = AddInstruction(OpCode.NOP);
+            using (InsertSequencePoint(syntax.Condition))
+                ConvertExpression(model, syntax.Condition);
+            Jump(OpCode.JMPIF_L, startTarget);
+            breakTarget.Instruction = AddInstruction(OpCode.NOP);
+            PopContinueTarget();
+            PopBreakTarget();
+        }
+    }
+}
